// @types/types.d.ts
import {Collection, SlashCommandBuilder, CommandInteraction } from "discord.js";

declare global {
  namespace NodeJS {
    interface ProcessEnv {
      CLIENT_ID: string,
      TOKEN: string,
    }
  }
}

declare module "discord.js" {
  export interface Client {
    slashCommands: Collection<string, SlashCommand>
  }
}

export interface BotEvent {
  name: string,
  once?: boolean | false,
  execute: (...arg) => void
}

export interface SlashCommand{
  name: string,
  data: SlashCommandBuilder | any, 
  async execute: (interaction: CommandInteraction) => Promise<void>
}

export {}